safari-600.7-branch/LayoutTests/js/dom/object-prototype-constructor-expected.txt

This is a test case for bug 3537.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS f.constructor is Foo.Bar
PASS typeof f.constructor is "function"
PASS f2.constructor is Foo2.Bar
PASS typeof f2.constructor is "function"
PASS f3.constructor is Foo3.Bar
PASS typeof f3.constructor is "function"
PASS successfullyParsed is true

TEST COMPLETE
